feat: add github config to dir target and move index.html to shared output
Add GitHubPagesConfig struct to dir target for .nojekyll/CNAME support.
Move index.html generation from github_pages target into write_common_files
so all targets produce a landing page.
